Move long running api test to misc test
