+++ tests/tt_metal/tt_metal/CMakeLists.txt
@@ -72,6 +72,7 @@ add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/perf_microbenchmark)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/noc)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/lightmetal)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/sfpi)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/misc)
 
 add_custom_target(
     metal_tests
@@ -89,6 +90,7 @@ add_custom_target(
         unit_tests_noc
         unit_tests_lightmetal
         unit_tests_sfpi
+        unit_tests_misc
 )
 
 add_library(metalium_test_files INTERFACE)
